containers: move getContainers and getContainerLockAndState to this new file
This will help to factorise code from several subcommands:
- list-containers (#162)
- status (#313)
- gc
